requirements.txt
openai==1.93.0
torch==2.3.0
transformers
chromadb==1.0.9
sqlglot==18.17.0
numpy
pandas
func_timeout
matplotlib
sentence_transformers==3.3.1
pysqlite3-binary
posthog==3.21.0
openpyxl
psutil
redis
redis[hiredis]
cassandra-driver
mistralai
datasketch
nltk
google-cloud-bigquery
snowflake
mysql-connector-python